+++ example/ios/Runner/GeneratedPluginRegistrant.m
@@ -3,8 +3,18 @@
 //
 
 #import "GeneratedPluginRegistrant.h"
+
+#if __has_include(<cloud_firestore/CloudFirestorePlugin.h>)
 #import <cloud_firestore/CloudFirestorePlugin.h>
+#else
+@import cloud_firestore;
+#endif
+
+#if __has_include(<firebase_core/FirebaseCorePlugin.h>)
 #import <firebase_core/FirebaseCorePlugin.h>
+#else
+@import firebase_core;
+#endif
 
 @implementation GeneratedPluginRegistrant