+++ lib/src/write_task.dart
@@ -1,0 +1,7 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+
+class WriteTask {
+  DocumentReference document;
+  Map<String, dynamic> data;
+  bool merge;
+}