+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.2.4
+
+- cleaned up the public documentation to expose only MockFirestoreInstance.
+
+## 0.2.3
+
+- fixed code health related issues.
+
+## 0.2.2
+
+- added support for isGreaterThanOrEqualTo, isLessThan, isLessThanOrEqualTo in CollectionReference.where
+- implemented mock DocumentReference.delete
+
 ## 0.2.1
 
 - Fixed snapshots not firing several times.